third_party/blink/web_tests/fast/dom/Document/CaretRangeFromPoint/caretRangeFromPoint-with-before-style-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS testIt("margin", -2) is "a"
PASS testIt("margin", -1) is "a"
PASS testIt("margin", 0) is "a"
PASS testIt("margin", 1) is "b"


PASS testIt("noMargin", -2) is "a"
PASS testIt("noMargin", -1) is "a"
PASS testIt("noMargin", 0) is "a"
PASS testIt("noMargin", 1) is "b"
